Add a useSection hook
