+++ apps/judicial-system/web/src/components/PageLayout/utils/index.ts
@@ -4,17 +4,7 @@ import {
   CaseState,
   CaseType,
   isInvestigationCase,
-  isRestrictionCase,
-  User,
 } from '@island.is/judicial-system/types'
-import {
-  getCourtSections,
-  getCustodyAndTravelBanProsecutorSection,
-  getExtenstionSections,
-  getInvestigationCaseCourtSections,
-  getInvestigationCaseExtenstionSections,
-  getInvestigationCaseProsecutorSection,
-} from '@island.is/judicial-system-web/src/utils/sections'
 
 interface TranslationStrings {
   dismissedTitle: string
@@ -72,49 +62,4 @@ export const caseResult = (
   } else {
     return 'Niðurstaða'
   }
-}
-
-export const getSections = (
-  translationStrings: TranslationStrings,
-  workingCase?: Case,
-  activeSubSection?: number,
-  user?: User,
-) => {
-  return [
-    isRestrictionCase(workingCase?.type)
-      ? getCustodyAndTravelBanProsecutorSection(
-          workingCase || ({} as Case),
-          activeSubSection,
-        )
-      : getInvestigationCaseProsecutorSection(
-          workingCase || ({} as Case),
-          activeSubSection,
-        ),
-    isRestrictionCase(workingCase?.type)
-      ? getCourtSections(workingCase || ({} as Case), user, activeSubSection)
-      : getInvestigationCaseCourtSections(
-          workingCase || ({} as Case),
-          user,
-          activeSubSection,
-        ),
-    {
-      name: caseResult(
-        { dismissedTitle: translationStrings.dismissedTitle },
-        workingCase,
-      ),
-    },
-    isRestrictionCase(workingCase?.type)
-      ? getExtenstionSections(workingCase || ({} as Case), activeSubSection)
-      : getInvestigationCaseExtenstionSections(
-          workingCase || ({} as Case),
-          activeSubSection,
-        ),
-    isRestrictionCase(workingCase?.type)
-      ? getCourtSections(workingCase || ({} as Case), user, activeSubSection)
-      : getInvestigationCaseCourtSections(
-          workingCase || ({} as Case),
-          user,
-          activeSubSection,
-        ),
-  ]
 }
